Refactor IT engine package (#8833)
